Revert the removal of the configuration patch
The patch is used to catch a corner case for building grconvert. The libtirpc is
required to build grconvert on RedHat-like distros only. When libtirpc is
missing, the build proceeds without building grconvert, which is then caught by
the sanity checks. Other systems can build grconvert with the libraries
provided.

If we just add libtirpc in LIBS, then systems that do not need libtirpc in order
to build grconvert will fail at linking, even though they can actually build
grconvert.